Badge and cover now optional downloads

+++ tasks.py
@@ -4,6 +4,7 @@ from datetime import datetime
 import shutil
 import json
 import os
+from os.path import exists
 
 user = ret_auth.user
 key = ret_auth.key
@@ -32,20 +33,30 @@ else:
     with open("cache.json", "wb") as f:
         f.write(response.content)
     data = response.json()
+
+cover_path = 'static\\images\\cover.png'
+badge_path = 'static\\images\\badge.png'
+
+cover_exists = exists(cover_path)
+badge_exists = exists(badge_path)
+
+if cover_exists == True:
+    pass
+else:
     game_cover = data[0]['GameIcon']
     cover_response = requests.get(f'https://i.retroachievements.org/{game_cover}', stream=True)
-    cover_path = 'static\\images\\cover.png'
     cover = open(f'{folder_name}\\cover.png','wb')
     shutil.copyfileobj(cover_response.raw, cover)
     cover.close()
+if badge_exists == True:   
+    pass
+else: 
     ach_url = data[0]['BadgeURL']
     ach_response = requests.get(f'https://i.retroachievements.org/{ach_url}', stream=True)
-    badge_path = 'static\\images\\badge.png'
     ach = open(f'{folder_name}\\badge.png','wb')
     shutil.copyfileobj(ach_response.raw, ach)
     ach.close()
     
 game_title = data[0]['GameTitle']
-
 ach_title = data[0]['Title']
 ach_desc = data[0]['Description']
